FogBugz #338696: Create recipe to build PCIe benchmarking module and apps
This is to create the recipes to build PCIe module and applications in
Angstrom. Also creates a new pcie-console-image for user to build the
image that includes the performance benchmarking application and module.

Currently GSRD is using kernel 3.10-ltsi, and PCIe reference design is
using kernel 4.1, and there are kernel modules in GSRD that uses features
in kernel 3.10-ltsi, but is removed in kernel 4.1, therefore the RFS image
is split into two. These images will be merged into a single image,
potentially called altera-refdesign-image.

Signed-off-by: Tien Hock Loh <thloh@altera.com>