FogBugz #426643: Fallback static IP for GSRD
This patch fixes an issue where the devkit is connected to a switch that doesn't
have DHCP server, or is directly connected to a PC without a DHCP server
running, the LCD will print "No IP Obtained" instead of a local-link IP address.
With a delay > 60 seconds, connman will fallover to ipv4ll and assign
a local link 169.254.x.x address, and only then the script will write the IP
address to the LCD display.

Signed-off-by: Tien Hock Loh <thloh@altera.com>