use require instead of include
The include directive does not produce an error when the file cannot
be found. Consequently, it is recommended that if the file you are
including is expected to exist, you should use require instead of
include. Doing so makes sure that an error is produced if the file
cannot be found.

Signed-off-by: Chang Rebecca Swee Fun <rebecca.swee.fun.chang@intel.com>